[#41] Refactor : redux 모듈 패턴 변경
1. 같은 데이터 구조를 가져오는 경우, 각 기능별 모듈에 상태를 따로 두고, 각각의 상황에서 요청하여 응답 받은 데이터를 저장하도록 합니다.
- 기존 프로필 페이지에서 사용자 작성 게시글 조회를 articles 모듈에서 담당하여 api 요청/응답 받던 것을 profile 모듈으로 이동하였습니다.
2. 전반적인 redux 모듈 패턴은 '기능 별로' 디렉토리를 나누어 관리하고 Actions, Reducer, Saga 파일을 따로 둡니다.
- 기존 Types 파일에서 선언하던 액션 타입은 Actions 파일에 함께 작성하도록 합니다.
