+++ frontend/src/modules/articles/articlesReducer.js
@@ -8,11 +8,9 @@ import {
 	READ_ARTICLE_SUCCESS,
 	READ_MY_ARTICLES_FAILURE,
 	READ_MY_ARTICLES_SUCCESS,
-	READ_USER_ARTICLES_FAILURE,
-	READ_USER_ARTICLES_SUCCESS,
 	UNLOAD_ARTICLE,
 	UNLOAD_ARTICLES,
-} from './articlesTypes';
+} from './articlesActions.js';
 
 const initialState = {
 	articles: {
@@ -47,18 +45,6 @@ const articles = handleActions(
 			articles,
 		}),
 		[READ_ALL_ARTICLES_FAILURE]: (state, { payload: error }) => ({
-			...state,
-			error,
-		}),
-		[READ_USER_ARTICLES_SUCCESS]: (state, { payload: newArticles }) => ({
-			...state,
-			articles: {
-				totalCount: newArticles.totalCount,
-				count: state.articles.count + newArticles.count,
-				data: [...state.articles.data, ...newArticles.data],
-			},
-		}),
-		[READ_USER_ARTICLES_FAILURE]: (state, { payload: error }) => ({
 			...state,
 			error,
 		}),
