+++ frontend/src/modules/editor/articleEditorActions.js
@@ -1,11 +1,13 @@
 import { createAction } from 'redux-actions';
-import {
-	CHANGE_FIELD,
-	INITIALIZE,
-	MODIFY_ARTICLE,
-	SET_ORIGINAL_ARTICLE,
-	WRITE_ARTICLE,
-} from './articleEditorTypes';
+import { createRequestActionTypes } from '../../services/createRequestSaga';
+
+export const INITIALIZE = 'articleEditor/INITIALIZE';
+export const CHANGE_FIELD = 'articleEditor/CHANGE_FIELD';
+export const [WRITE_ARTICLE, WRITE_ARTICLE_SUCCESS, WRITE_ARTICLE_FAILUER] =
+	createRequestActionTypes('articleEditor/WRITE_ARTICLE');
+export const SET_ORIGINAL_ARTICLE = 'articleEditor/SET_ORIGINAL_ARTICLE';
+export const [MODIFY_ARTICLE, MODIFY_ARTICLE_SUCCESS, MODIFY_ARTICLE_FAILURE] =
+	createRequestActionTypes('articles/MODIFY_ARTICLE');
 
 export const initialize = createAction(INITIALIZE);
 export const changeField = createAction(CHANGE_FIELD, ({ key, value }) => ({
