+++ frontend/src/modules/profile/profileReducer.js
@@ -1,16 +1,24 @@
 import { handleActions } from 'redux-actions';
 import {
 	INITIALIZE_PROFILE,
+	READ_MEMBER_ARTICLES_FAILURE,
+	READ_MEMBER_ARTICLES_SUCCESS,
 	READ_PROFILE_FAILURE,
 	READ_PROFILE_SUCCESS,
+	UNLOAD_ARTICLES,
 	UPDATE_PROFILE_FAILURE,
 	UPDATE_PROFILE_IMAGE_FAILURE,
 	UPDATE_PROFILE_IMAGE_SUCCESS,
 	UPDATE_PROFILE_SUCCESS,
-} from './profileTypes';
+} from './profileActions';
 
 const initialState = {
 	profile: null,
+	articles: {
+		totalCount: null,
+		count: null,
+		data: [],
+	},
 	updatedProfile: null,
 	error: null,
 };
@@ -23,6 +31,18 @@ const profile = handleActions(
 			profile,
 		}),
 		[READ_PROFILE_FAILURE]: (state, { payload: error }) => ({
+			...state,
+			error,
+		}),
+		[READ_MEMBER_ARTICLES_SUCCESS]: (state, { payload: newArticles }) => ({
+			...state,
+			articles: {
+				totalCount: newArticles.totalCount,
+				count: state.articles.count + newArticles.count,
+				data: [...state.articles.data, ...newArticles.data],
+			},
+		}),
+		[READ_MEMBER_ARTICLES_FAILURE]: (state, { payload: error }) => ({
 			...state,
 			error,
 		}),
@@ -41,6 +61,14 @@ const profile = handleActions(
 		[UPDATE_PROFILE_IMAGE_FAILURE]: (state, { payload: error }) => ({
 			...state,
 			error,
+		}),
+		[UNLOAD_ARTICLES]: (state) => ({
+			...state,
+			articles: {
+				totalCount: null,
+				count: null,
+				data: [],
+			},
 		}),
 	},
 	initialState,
