+++ frontend/src/modules/user/userActions.js
@@ -1,0 +1,10 @@
+import { createAction } from 'redux-actions';
+import { createRequestActionTypes } from '../../services/createRequestSaga';
+
+export const TEMP_SET_USER = 'user/TEMP_SET_USER'; // 새로고침 이후 임시 로그인 처리
+export const [CHECK, CHECK_SUCCESS, CHECK_FAILURE] = createRequestActionTypes('user/CHECK');
+export const LOGOUT = 'user/LOGOUT';
+
+export const tempSetUser = createAction(TEMP_SET_USER, (user) => user);
+export const check = createAction(CHECK);
+export const logout = createAction(LOGOUT);
